cpu-o3: Extract trace-mode logic from Commit
Why:
- Keep Commit mainline focused; isolate trace-mode behavior.

What:
- Move trace-mode helpers and CommitTrace logging into src/cpu/o3/trace/CommitTrace.cc.
- Keep commit.cc as thin delegation points.

Verification:
- Trace regression (short list, warmup/sample=50000): baseline vs refactor has no diffs in committedInsts/ipc/simTicks/finalTick.
